fix: prevent app from crashing when a locale is missing
This adds a check when setting locale: if locale is not present, the default locale is set.
Unchecked, missing locale could come from autodetection or from local storage.

This also disables autodetection. Users should, hopefully, find the language switcher on top of the page if they need locale to be changed.