refactor(cli): merge the algorithm steps into analysis step
The only part that is not moved is the tree attachment, because it is serial.

This move seems to be breaking the tree attachment, probably because the matches returned from finding nearest nodes are pointing to the copy of the tree made when crossed the webworker process boundary. This is to be investigated and fixed.

This also break the application (algorithm saga) and this also have to be fixed.