refactor(cli): make sure new nodes are attached to the tree again
Now that the `treeFindNearestNodes()` is running inside webworker (and copied using structured cloning algorithm), the `match` node "pointers" are no longer preserved and we cannot attach anything to them.

To mitigate this, we introduce explicit numeric ids to the reference nodes, so that we can the find these on the tree again, when attaching. This is a bit less efficient, but keeps everything in one webworker and makes code cleaner.