chore: use webpack v5
Upgrades webpack to version 5.0.0 and Next.js to version 9.4.5.

Webpack 5.0.0 was released a couple of days ago.
https://webpack.js.org/blog/2020-10-10-webpack-5-release/

The primary motivation for this instant upgrade is much easier integration of Webassembly with Webpack 5 compared to Webpack 4, which is very handy now that we are planning to reimplement the core algorithms in C++.

Transition is surprisingly smooth so far.


Notable changes:

 - yarn resolution for webpack is updated to `5.0.0`, so that Next.js picks it up automagically.

 - Webpack now supports webworkers natively, so we dropped `threads-plugin` entirely. The new `Worker` syntax seems to be functioning with `threads.js` out of the box, including worker pools. We use global `Worker`, instead of the `Worker` previously imported from `threads.js`.

 - We temporarily use master branch of `next-transpile-modules`, which was updated to webpack 5 but have not yet been published on npm.