feat!: remove clade markers from Gene Map, remove clades.json
BREAKING CHANGE: This removes `virus.json` file and modifies `Virus` data structure.

For a while now, we assign clades by looking at the closest node in the tree, and we don't rely algorithmically on clade definitions in `clades.json`. The only place where clade definitions were used is displaying clade-defining mutation markers on Gene Map.

We agreed that with introduction of subclades we want to remove `clades.json`, rather than update it. It also means removing clade-defining mutation markers on Gene Map in the web app. This commit makes this happen.

Affected are:
 - Gene Map in the web app does not show clade markers anymore
 - CLI users don't have to provide `clades.json` for custom viruses