fix: ensure auspice strings interpolate correctly
Our recent upgrade to "i18next": "19.8.3" and   "react-i18next": "11.7.3" broke some of the localized strings in auspice parts of the application. Apparently, the cause of the breakage is a mismatch of versions of this packages used in nextclade and in auspice.

This PR forcefully pins versions of these packages (via yarn resolutions), so that auspice uses nexclade's versions of these packages. This fixes string interpolation.