refactor: use ghc-filesystem as a replacement for std::filesystem
This switches nextalign CLI to use https://github.com/gulrak/filesystem as a drop-in replacement for `std::filesystem`.

Despite having C++17 enabled, there is a complication with Applce clang and macOS target version: we want to target older macOS versions (down to 10.12) which don't ship some of the features we need.

We solve the problem by using the replacement library.

There are also some minor formatting issues fixed.