feat: add an option to allow out-of-order writes
This adds `--out-of-order` CLI option which allows parallel processing out of order.

Without this flag the program will wait for all results from the preceding sequences to be written to the output files before writing to the results for the next sequence. 
This might introduce waiting times, but ensures that the resulting sequences are written in the same order as they occur in the inputs.

Specifying this flag allows to lift this requirement, which might make the processing faster, by eliminating waits, but might also lead to results written out of order. In this case, the order of results is not specified and depends on thread scheduling and processing times of individual sequences.