feat: reimplement aminoacid change detection based on extracted peptides
This reimplements aminoacid changes search (substitutions and deletions). We now  compare reference and query peptides (extracted by the preceding call to Nextalign), one aminoacid at at time, and deduce changes. We then report the change and relevant nucleotide context surrounding this change.

Previously we reported one-to-one mapping of aminoacid changes to corresponding nucleotide changes. However, it was not always accurate, because if there are multiple nucleotide changes in a codon, the direct correspondence might not always be established without knowing the order in which nucleotide changes have occurred. And in the context of Nextclade we don't have this information.