fix: don't store sequences in memory beyond the scope where they needed
This groups the analysis results to a struct (`AnalysisResult`) and uses it as a field in the overall results struct (`NextcladeResult`).

This way, we can easily save analysis results outside of "main loop" for writing results output files and tree node attachment, while the sequences will go out of scope and free up memory (or rather not leak it).