feat: improve cli console logging
This improves console logging messages of Nextclade CLi and Nextalign CLI.
The messages are not prefixed with message's log level (verbosity level) and the name of the application. Some of the error and warning messages have been reworded to better convey their meaning (but there are still some work to do here).

This should improve log readability and facilitate debugging when using Nextclade CLi and Nextalign CLI in complex pipelines.

Examples:

```
[ INFO] Nextclade: |   302 | USA/UT-01231/2020                        |   89545 |      48 |   29903 |       0 |
[ WARN] Nextclade: Warning: in sequence "USA/UT-01231/2020": When processing gene "ORF7b": Unable to align: no seed matches. Note that this gene will not be included in the results of the sequence
```

```
[ERROR] Nextclade: Error: when running the internal parallel pipeline: When parsing input sequences: Input stream ("path/to/an_empty_file.fasta") is empty or corrupted. Aborting.
```