feat: distinguish between fatal and non-fatal errors
We want to handle some of the errors during the analysis of sequences such that it only interrupts the analysis of the affected sequence. Let's call it "non-fatal error" and inherit corresponding exception classes from `ErrorNonFatal` and let's inherit the remaining exception classes from `ErrorFatal`. This way we can catch these separately and handle differently.