perf: remove exceptions from nextalign algo, disable exception catching
We've been catching some of the exceptions and turn them into warning messages, in order to not crash the processing entirely when it is recoverable (for example if one of gene translations throws an exception, the rest of the genes should still be processed)

if explicitly enabled, exception catching is very slow in WebAssembly, so we cannot enable it. And if disabled (which is the default), the catch clauses won't work and the web app will show incorrect results (errors where there should not be). 

Removing exceptions and replacing them with explicitly returned status codes fixes this problem.

However, now the code combines exceptions with status codes inconsistently, which is ugly and error prone. But I don't see any other way as of right now.