feat: link aminoacid and nucleotide changes by position
Here we establish a many-to-many relation between aminoacid changes (substitutions, deletions) and nucleotide changes (substitutions, deletions).

We group changes by positions:
 - nucleotide substitution is related to aminoacid change iff nucleotide position is within the codon nucleotide range of that aminoacid
 - nucleotide deletion is related to aminoacid change iff nucleotide deletion range intersects the codon nucleotide range of that aminoacid

We only state the fact of spatial locality of these changes and we say nothing about the causal relations between nucleotide changes and aminoacid changes, as this can not always be easily deduced from sequences only.

This data is automatically propagated to JSON result output and to the web app (but is not used there yet)