fix: ensure non-fatal errors don't crash the app
Resolves #434

The exception catching flag was not properly concatenated to the list of compiler flags in cmake file (I can never remember all the string/list cmake concat syntax rules). Therefore, exception catching in the wasm module was not properly re-enabled.

As a result, non-fatal exceptions (such as sequence alignment failures) were not caught on the cpp-to-js boundary. They were crashing the entire wasm module and reported to users in popup windows, as fatal. Sequence analysis was also halting without indication in the results table.

This PR fixes the string concat in cmake, correcting the entire chain of these events.
The non-fatal errors are now reported in the table.