fix(cli): make sure subcommands don't require root args
HACK: Prepends "run" subcommand to the command line, to be able to use subcommands (`dataset fetch`,
`dataset list`) while preserving the user interface and not introducing breaking changes.

Reason: There seem to be no way of stopping required flags of the root command to propagate to subcommands,
so for example `dataset fetch` subcommand will fail if `--input-fasta` is required in the root. We moved
`--input-fasta` to the "run" subcommand. However, in order to preserve the old interface we now need to fake the
"run" command when it's not there. Without it users would have to prepend "run" subcommand to all Nextclade
invocations in the existing scripts and pipelines. Might be dropped in the next major release.