chore: use libmusl for cross-distro static linux builds
Networking C++ libraries introduced in previous commits use glibc and system calls in order to provide DNS resilution and network I/O, which can vary between Linux distributions, kernel versions and glibc versions.

This means that even static executable, with all dependencies built as static libraries rely on the presence of a particular version of glibc on the target machine. If there's mismatch. The console log is then spiced up with numerous linker warnings, because it knows we are producing binaries that are not portable.

This commit introduces:
 - libmusl as a small static glibc alternative, which does not rely on Linux distribution specifics
 - a libmusl-based GCC 9 cross-toolchain, including libstdc++ that is linked against libmusl, to ensure every dependency, recursively, uses libmusl
 - locally built conan packages for openssl, libcurl, c-ares, poco, tbb with small patches for proper resolution of interdependencies and to ensure fully-static build