fix: make sure release executables work and are fully static
Previously, dependency were built with a mixture of glibc and musl, due to system headers, libs and executables leaking into the build. This was causing random  crashes in release mode.

Here we ensure that only musl is used.

We patch up the symbols used in gtest but missing in musl. Test executables still seem to be broken in release mode. They cannot be executed for some reason. This needs to be investigated. Disabled tests in release mode for now.