feat: use jemalloc
jemalloc is an alternative implementation of malloc:

After moving to musl implementation of libc, Nextclade has become about 10-12 times slower due to musl malloc implementation being simpler and slower than that of glibc.

Now we link agaginst jemalloc, which has a faster implementation. We are now  back to the almost the speeds of glibc: only 0-10% slower depending on inputs.

See: http://jemalloc.net/