feat(nextalign): optionally stop translation on first stop codon
The new Nextalign option `translatePastStop` (`false` by default) controls whether translation proceeds past stop codons.

If `translatePastStop` is `false`, the peptide past fist stop codon is filled with gaps.
If `translatePastStop` is `true`, the stop codons are ignored and there might be multiple stop codons in the resulting peptide.

Nextclade uses `translatePastStop=true` to preserve current behavior.

We might or might not want to expose this parameter in CLI.