fix: ensure there are no false QC schema version warnings
Currently Nextclade expects QC config `schemaVersion` to match Nextclade version itself. This means QC `schemaVersion` needs to be updated on every release.

This is no longer convenient now that we will be transitioning to the separate datasets repo. So instead we freeze the QC `schemaVersion` on a particular semver string hardcoded in Nextclade code.

Analysis results JSON `schemaVersion` was also refactored to bring the two versions closer together, even though Analysis results JSON `schemaVersion` is only used to write results and not checked or read anywhere for now.