fix: protect only first codon before frame shift detection
So far we gap-stripped genes *after* frame shift detection. Right before stripping we would also protect the first and last not-all-gaps codons by replacing gap character `-` with `N` in them, in order to preserve reading frame during subsequent gap stripping. So the old flow was:
 - detect frame shifts
 - protect first and last codons
 - strip gaps

This was incorrect, because the frame shift detector have treated `-` characters in unprotected partial codons as deletions, which caused false-positive detections of frame shifts in the beginning. In reality these are just incomplete sequences and the gaps were introduced by the alignment.

This commit splits the protection code and stripping code and puts protection before frame shift detection. The new flow is:
 - protect first codon (and not last)
 - detect frame shifts
 - strip gaps
This way the false-positives are eliminated.

We no longer need to protect the last non-gap codon. Translation can now handle the genes with length that is not divisible by 3, so we no longer need to worry about the completeness of the tail codon.