fix: translate past first stop codon in nextclade web
After #504 when translation past stop codons became optional, Nextalign and Nextclade CLI gained their flags to toggle this option. However, Nextclade Web was incorrectly using the default value, and translation was terminated on the first stop. 

This fixes the option's value so that Nextclade Web translates full peptides as expected.