feat(web): truncate peptide context and list of changes in tooltips
Currently when a large group of consecutive mutations or deletions appear in a peptide, mouse hovering this block displays a tooltip which may not fit to the screen due to both the width of the peptide context table and height of the list of mutations and deletions.

This PR truncates the context table and the lists, both in the middle, keeping the edges, all making sure tooltips are not too large and that they fit to the screen nicely.