feat: attach temporary aa mutation maps to the tree nodes
This prepares the nodes for search for private aa mutations.
Similarly to how it's done nuc mutations, this search is performed concurrently, but the map construction is a sequential operation, so we perform it during the "tree preprocessing" step and then attach the results to the nodes. The "find nearest node" operation will find use this information to find private aa mutations concurrently (to be implemented). The maps are removed in the "tree postprocessing" step to that they don't appear on the output tree.