fix: revert: only consider fully static build on linux 2fcb478
This reverts https://github.com/nextstrain/nextclade/commit/2fcb4781ccb1a6c24b49661b865eaf1d04a3f899, which wrongly restricted static linking of dependencies to Linux builds only. As dependency dynamic libraries were now required on macOS, but not redistributed along with the executables, Nextclade was crashing on users' machines, in particular trying to find a dynamic library for Intel TBB.

It went unnoticed during dev and on ci, because these libraries are present in the build directory after the build.

This PR changes build such that macOS binaries are also statically linked and should be self-contained again now.

We might want to implement a smoke test in a clean environment with no build artifacts present other than the executables, to verify that they are truly static and self-contained. This is however challenging for macOS builds, as there are no containers and VMs are quite expensive and clunky to deal with.

Resolves https://github.com/nextstrain/nextclade/issues/595