fix: avoid race between data fetch and read when url params are used
When datasets feature was added, a data race was introduced due to lack of synchronization between asynchronous data fetching from locations provided by URL params and reading this data before launching the algorithm.

The order of operation was as follows:
1. Retrieve URLs from URL params
2. Start fetching from these urls
3. Save fetched data in redux store upon arrival
4. Read data from redux store. If it's empty, load default data from a dataset.
5. Do some other operations.
6. Launch the algorithm.

However, step 4 was not waiting until steps 3 is completed. So it could happen that the data read yielded empty, because fetching was still in progress. In this case the  algorithm was started with the default data, even though URL params are provided.

In dev mode, with the data coming from a local server that was working, because data fetch was fast and completed before the algorithm was started. However when data is served from a slower external location that was no longer the case.

In order to avoid the race, this PR changes the order of operations, and merges data  fetch and data reads, so that they are an atomic operation:
1. Retrieve URLs from URL params
2. Save URLs data in redux store
3. Fetch the data, if it's empty, load default data from a dataset.
4. Do some other operations.
5. Launch the algorithm.

This avoids the race condition and ensures that the data is fetched before its used.