fix: reduce number of webworker timeout errors
On a busy computer/browser Nextclade Web sometimes crashes with a webworker timeout message. This is due to either webworker or the main thread being busy for an extended period of time, leading to communication delays between them. This can happen if a large set of sequences is analyzed on an underpowered machine, or on a machine with most of the resources (CPU, RAM) already occupied by other  browser tabs and processes.

Additionally, the wasm modules for workers are lazy-loaded from the server, so if there are network problems or if network is too slow, the initialization of a new worker, might be additionally delayed by the network overhead. In particular, after fresh releases, when these modules are not yet in the browser's cache.

Here we increase the timeout, so that there is some more room for webworkers to communicate. The default timeout is 10 seconds, now it's 60 seconds.

A better fix would be:
 - to reduce number of workers to be initialized, by reusing existing workers and wasm modules
 - to reduce number and volume of communication required between workers and the main thread (by structuring function calls better and using more efficient dafa formats for messages)
 - to preemptively initialize workers in background during idle periods