fix: don't add private reversions when query aa is unknown
Here we filter out substitutions and deletions from results, where positions fall into ranges with aminoacid X in query peptide.

In these cases the information is simply missing from the query and in the `processNodeMutations()` above if a substitution or a deletion is missing from query, but present in the node, a reversion is added to the results (a mutation from the node all the way to what was in the ref seq). We chose to remove these,
because we believe that a reversion is a less likely event and most of these are the sequencing errors followed by uncertainty in translation.