fix: error when `--genes` flag contains a subset of genes
When `--genes` flag was provided to Nextclade CLI, which contains a list with only a subset of genes present in gene map, it would result i an error "Reference peptide not found".

This was caused by the incorrect assumption in this part of code, that all ref genes are always present, which is not the case when genes are excluded with some of the genes omitted in the `--genes` flag.

Here  I made it so that if a reference peptide is not found, we now simply continue iterating, ignoring that gene.