fix: undefined behavior in tree json acccess
In case where the key is not present the const version of the get function:
```
const json& get(...) const
```
was returning `defaultValue` which is a reference to a temporary, and hence was destroyed after return, thus the returned reference was dangling.

Here I remove the get function, and move conditionals to the caller. I also simplify the chained access with a json pointer.