fix: enforce precondition on stripped sequences, replace invalid test
The test `ReportsAlignmentStartAndEnd`
https://github.com/nextstrain/nextclade/blob/b007490e0107a07b65534b175a4ffd6350685963/packages/nextclade/src/__tests__/findNucChanges.test.cpp#L22-L26
was crashing here
https://github.com/nextstrain/nextclade/blob/b007490e0107a07b65534b175a4ffd6350685963/packages/nextclade/src/analyze/findNucChanges.cpp#L40
accesing string `refStripped` at index 3 at while it's  length is 3.

This is caused by the fact that the function assumes that query and ref sequences have the same length and cab be indexed in unison, but in the test they were different length.

This should have never happened, as the function is meant to be used on stripped sequences.

We don't want to add any runtime checks for performance reasons, and Nextclade only has access to stripped sequences as of now. But I added a precondition to check is query and ref lengths are equal.

I also replaced the invalid test with what Richard suggested, and that test resembles the reality a bit more.

Also added a doc comment to the function, because why not.