test: add tests for trenslateGenes() where peptides are very small
Followup of https://github.com/nextstrain/nextclade/pull/674, https://github.com/nextstrain/nextclade/pull/683

Adds test cases for gene lengths 0, 1, 2 and 3 nucleotides. It's expected that the cases with 0, 1 and 2 nucs will result in an empty peptide and it will not be emitted, and for case of 3 nucs, a 1-AA peptide is emitted.