fix: exit with correct exit code on exceptions
It could be that some of the exceptions that occurred in Nextclade and Nextalign libraries were handled, but then the program execution would end with the default implicit 0 exit code returned from main. This was due to a redundant forgotten extra try-catch block.

This removes this try-catch block, so that the exceptions are routed to the correct outer catch clause, which exits with code 1.