fix: account for ambiguous nucs when matching private mut labels
https://github.com/nextstrain/nextclade/issues/698

This replaces strict equality of nucleotides with the equivalence when searching for private mutation labels in the label map.

I replaced `operator==(nuc1, nuc1)` call with `isMatch(nuc1, nuc2)` call, which matches both nucleotides against IUPAC table of equivalence. This means that ambiguous nucleotides will be considered equivalent to their corresponding non-ambuguous ones. Nucleotide `N` will be equivalent to any nucleotide, which may or may not be what we need.