fix(web): center markers in sequence views
Resolves https://github.com/nextstrain/nextclade/issues/708 https://github.com/nextstrain/nextclade/issues/631
Followup of https://github.com/nextstrain/nextclade/pull/638, https://github.com/nextstrain/nextclade/pull/644

Continues on the adventure of centering various markers in sequence views, started in https://github.com/nextstrain/nextclade/pull/644. 

The goal is to make the different markers to be aligned to their positions in the sequence properly, as well as to be aligned between each other. The challenge is that there's enough space to display them and that some of the markers are single positions (substitutions), while others are ranges (Ns), and then can overlap, hide each other and cause all kinds of visual anomalies.

This time I adjusted markers:
 - for all ranges (peptide fs, nuc fs, aa Xs, nuc Ns, nuc dels, unsequenced nucs)
 - for nucleotide substitutions (they are non-range, 1-nuc markers)
by subtracting  half of a width of the nuc/aa from their left positions, so that they are shifted left by a half a nuc/aa, ensuring they are anchored at the center of the first nuc/aa in the range. It is important to note that the width of a nuc/aa is variable, so the max of pixel-per-item and min-item-width need to be considered as a width.

We need to check how it plays with the AA mutation groups (which are the single-aa markers aggregated into groups), which has a more complex formula for positions (see https://github.com/nextstrain/nextclade/pull/638, https://github.com/nextstrain/nextclade/pull/644).