fix: exclude reversions of dels from consideration in "SNP clusters" QC
Resolves https://github.com/nextstrain/nextclade/issues/707
Related to https://github.com/nextstrain/nextclade/pull/689

This tries to address the fact that reversions of deletions are now included into "all private mutations" array and so are processed in the "SNP clusters" QC rule.
Because sometimes reversions of deletions can come in large contiguous chunks, these mutations contribute to the cluster counts, making QC scores for some of the sequences worse. This was unexpected.

In order to mitigate this, it was decided in the team chat to exclude reversions of deletions from the consideration in the "SNP clusters" QC rule.

The technical implementation is to remove the intersection of "all private substotutions" and "reversions of deletions" from "all private substitutions". Implemented rather naively, because these arrays are small and don't worth the hassle of of using any more advanced algorithms.

We may or may not to reconsider this. One thing to try could be to treat reversions of deletions as ranges. This would require introducing a concept of ranges of substitutions (similar to the currently existing ranges of deletions) or to treat reversions of deletions as insertions. The QC rule then could be adjusted to count these ranges and/or to have a lower weight. In any case it is substantially more work and potential breaking changes to file formats, compared to just excluding them. And because the "SNP clusters" rule have somewhat less of importance nowadays, it was decided to go the easy path for now and to focus on other, more important issues.