fix: shift sequence views such that centered markers are visible
Addresses some of the concerns for https://github.com/nextstrain/nextclade/issues/712

Here I shifted sequence (peptide) view's SVG viewBox 1/n of nucleotide (peptide) width to the right. This allows for the markers close to the begging or end of the view to be visible.

This is required after adjustments made for centering markers in 
https://github.com/nextstrain/nextclade/pull/638 https://github.com/nextstrain/nextclade/pull/644 https://github.com/nextstrain/nextclade/pull/709