fix: don't add "truncated" suffix to inserted fragment
This was a leftover when the fragment were displayed as a plain string. 

At some point the fragment has become a component of it's own and the "truncated" text was displayed separately, but I forgot to remove the suffix.

This fix it by removing the concatenation of the suffix to the fragment.