fix(web): ensure correct rendering of mutation at position 0
Resolves https://github.com/nextstrain/nextclade/issues/745

There was a redundant null check in the NucleotideMutationBadge component that was causing the position of a mutation at first position to be ignored (the number 0 is considered nullish in js). This checks is redundant anyway, because the position is never null, so I removed it. AminoacidMutationBadge component does not have this check, so it should behave correctly.