chore(build): parch ghc-filesystem cmake file to make it detectable
It seems that conan have recently changed the rules of how dashes are handled in package names.

Previously ghc-filesystem would generate `FindghcFilesystem.cmake` and then `ghcFilesystem` can be used in cmake's `find_package()` call and `ghcFilesystem::ghcFilesystem` as a lib dependency. However, now conan generates `Findghc-filesystem.cmake` and the build breaks because ghcFilesystem package cannot be found. Renaming this file is an easy fix.

We are about to replace all this nonsense with Rust, so I am not too worried about the monkey patching here.