fix(web): remove 'debug' package from production bundle
`debug` package uses `windows` global, which is not
present in WebWorker environment, causing undefined error in production.

Previously, the correct setup of the environment was handled by threads.js webpack plugin, but it currently does not work with Webpack 5, so we have to improvise. Removing `debug` package resolves the error for now.