refactor(web): move more state away from Redux and into Recoil
This removes massive amounts of Redux state handling code and replaces it with Recoil. Because some of the Recoil atoms are not yet initialized, certain components are in permanent loading state (due to React's Suspense). Next goal is feature completeness of Recoil state handling, compared to Redux, so that all components can render properly