feat(web): limit number of markers that can be displayed in seq views
This add a setting that defined maximum number of markers (mutations, deletions, missing, frame shifts) that can be displayed in nucleotide sequence views.

If this number is exceeded, a message is displayed instead.

The value is 300 by default and can be changed in settings. Value is persisted in local storage.