feat(web): limit number of aa mut badges in nuc gap tooltips
This limits number of aminoacid mutation badges in the tooltip of the nucleotide gap marker in sequence view.

This avoid screen overflow when there's too many mutations.

This happens with aa deletions mostly, so ideally we need to group aa deletions into ranges. But it have negative impact to other places which need access to individual deletions and is more work overall. Maybe later.