refactor: move variables into the function to avoid extra arguments
This branch is based on https://github.com/nextstrain/nextclade/pull/815

This moves 2 variables along with the computation logic into the function, closer to the usage. These variables were not used anywhere outside the function.

Helped me to better understand the changes in https://github.com/nextstrain/nextclade/pull/815, and can perhaps allow to simplify logic in it.