feat(web): add global limit on total number of displayed markers
We currently have a per-sequence limit for number of displayed sequence view markers. This works well if one or a few among sequences have many markers. Users can tune the threshold.

However, if different sequences have very different number of markers (it's not uncommon for a large virus that number of mutations can differ 10- or 100-fold across the batch of sequences), then it's difficult to find a middle ground.

If a limit was set too high browser can hang during rendering or kill the tab.

Here introduce a global threshold on markers - which is a sum of numbers from all sequences. Upon crossing the threshold no markers are display for any of the sequences. This is a more reliable safeguard against hangups and crashes.