fix(web): make 'invalid range' warning condition more optimistic
This avoids a warning when the range is empty as we now consider an empty range  a valid case.

Notably, we've encountered this when a frame shift happens to be due to a 1-base deletion of a stop codon. In this case we still want to report a frame shift situation, even if the frame shift range is technically empty.