fix(web): ensure init errors are not hidden
Nextclade Web has been hiding some of the errors that occur during initialization. Notably, if dataset server is not reachable or dataset index fetch fails for any reason, then Nextclade would just show loading spinner indefinitely.

This PR ensures that the error is properly handled and that an error message is shown in these cases.