feat(cli): allow to replace unknown nucs with Ns
This adds a flag `--replace-unknown` to `run` command, which allows to replace unknown nucleotide characters with 'N'.

By default, the sequences containing unknown nucleotide nucleotide characters are skipped with a warning - they
are not aligned and not included into results. If this flag is provided, then before the alignment,
all unknown characters are replaced with 'N'. This replacement allows to align and analyze these sequences.

The following characters are considered known:

```
-ABCDGHKMNRSTVWY
```