feat(cli): organize verbosity flags towards the end of help message
Currently --verbosity, --silent, -v and -q args are missorted in the --help message text.

Here I inline the `clap-verbosity-flag` crate (only 1 file) and modify it, adding our custom flags and `display_order` annotations, such that the args are shown in the very end of the message, just before the `--help` arg.